Log stacktrace of work unit thread if it is slow to stop
While waiting for the work unit to stop, the stacktrace of its thread will be logged every 10s if it has changed since the last time it was logged.
This will allow users to more easily identify the problematic code that is not responding appropriately to interrupts, in some cases.